Refactorings to pull stuff out of FPComputeGraphQL
That file has simply gotten unwieldy, so lifted some bits out.
