+++ querki/scalajvm/app/querki/graphql/package.scala
@@ -1,0 +1,9 @@
+package querki
+
+import cats.effect.IO
+import cats.data.{EitherT, NonEmptyChain, ValidatedNec}
+
+package object graphql {
+  type SyncRes[T] = ValidatedNec[GraphQLError, T]
+  type Res[T] = EitherT[IO, NonEmptyChain[GraphQLError], T]
+}
